@@ -17,11 +17,18 @@
     "email": "support@digitalbazaar.com"
   },
   "license": "BSD-3-Clause",
-  "main": "index.js",
-  "module": "main.js",
+  "main": "./index.cjs",
+  "type": "module",
+  "exports": {
+    "import": "./main.js",
+    "require": "./index.cjs"
+  },
+  "browser": {
+    "./tests/util.js": "./tests/util-browser.js"
+  },
   "files": [
     "baseN.js",
-    "index.js",
+    "index.cjs",
     "main.js"
   ],
   "dependencies": {
@@ -36,26 +43,24 @@
     "babel-loader": "^8.0.6",
     "chai": "^4.2.0",
     "chai-bytes": "^0.1.2",
-    "core-js": "^3.6.4",
-    "cross-env": "^6.0.3",
-    "eslint": "^6.8.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^8.0.1",
     "eslint-config-digitalbazaar": "^2.1.0",
-    "karma": "^4.4.1",
+    "karma": "^6.3.4",
     "karma-babel-preprocessor": "^8.0.1",
     "karma-chrome-launcher": "^3.1.0",
     "karma-edge-launcher": "^0.4.2",
-    "karma-firefox-launcher": "^1.3.0",
-    "karma-ie-launcher": "^1.0.0",
-    "karma-mocha": "^1.3.0",
+    "karma-firefox-launcher": "^2.1.1",
+    "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
     "karma-safari-launcher": "^1.0.0",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-tap-reporter": "0.0.6",
-    "karma-webpack": "^4.0.2",
-    "mocha": "^7.0.0",
+    "karma-webpack": "^5.0.0",
+    "mocha": "^9.1.3",
     "mocha-lcov-reporter": "^1.3.0",
     "nyc": "^15.0.0",
-    "webpack": "^4.41.5"
+    "webpack": "^5.59.0"
   },
   "nyc": {
     "exclude": [
@@ -67,13 +72,12 @@
     ]
   },
   "engines": {
-    "node": ">=8.3.0"
+    "node": ">=10"
   },
   "scripts": {
     "test": "npm run test-node",
     "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks -t 30000 -A -R ${REPORTER:-spec} tests/*.spec.js",
-    "test-karma": "karma start",
-    "benchmark": "node benchmark/benchmark.js",
+    "test-karma": "karma start karma.conf.cjs",
     "coverage": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text-summary npm run test-node",
     "coverage-report": "nyc report",
     "lint": "eslint ."
